+++ lib/src/base/mega_app.dart
@@ -1,100 +1,9 @@
 import 'package:flutter/cupertino.dart'
     show CupertinoPageRoute, DefaultCupertinoLocalizations;
 import 'package:flutter/material.dart' show DefaultMaterialLocalizations, Icons;
-import 'package:flutter/services.dart';
 import 'package:mega/mega.dart';
 
-/// An application that uses Mega design.
-/// The [MegaApp] configures the top-level [Navigator] to search for routes
-/// in the following order:
-///
-///  1. For the `/` route, the [home] property, if non-null, is used.
-///
-///  2. Otherwise, the [routes] table is used, if it has an entry for the route.
-///
-///  3. Otherwise, [onGenerateRoute] is called, if provided. It should return a
-///     non-null value for any _valid_ route not handled by [home] and [routes].
-///
-///  4. Finally if all else fails [onUnknownRoute] is called.
-///
-/// If a [Navigator] is created, at least one of these options must handle the
-/// `/` route, since it is used when an invalid [initialRoute] is specified on
-/// startup (e.g. by another application launching this one with an intent on
-/// Android; see [dart:ui.PlatformDispatcher.defaultRouteName]).
-///
-/// This widget also configures the observer of the top-level [Navigator] (if
-/// any) to perform [Hero] animations.
-///
-/// If [home], [routes], [onGenerateRoute], and [onUnknownRoute] are all null,
-/// and [builder] is not null, then no [Navigator] is created.
-///
-/// {@tool snippet}
-/// This example shows how to create a [MegaApp] that disables the "debug"
-/// banner with a [home] route that will be displayed when the app is launched.
-///
-/// ```dart
-/// MegaApp(
-///   home: MegaScaffold(
-///     headerBar: MegaHeaderBar(
-///       center: const Text('Home'),
-///     ),
-///   ),
-/// )
-/// ```
-/// {@end-tool}
-///
-/// {@tool snippet}
-/// This example shows how to create a [MegaApp] that uses the [routes]
-/// `Map` to define the "home" route and an "about" route.
-///
-/// ```dart
-/// MegaApp(
-///   routes: <String, WidgetBuilder>{
-///     '/': (BuildContext context) {
-///       return MegaScaffold(
-///         headerBar: MegaHeaderBar(
-///           center: const Text('Home Route'),
-///         ),
-///       );
-///     },
-///     '/about': (BuildContext context) {
-///       return MegaScaffold(
-///         headerBar: MegaHeaderBar(
-///           center: const Text('About Route'),
-///         ),
-///       );
-///      }
-///    },
-/// )
-/// ```
-/// {@end-tool}
-///
-/// {@tool snippet}
-/// This example shows how to create a [MegaApp] that defines a [theme] that
-/// will be used in the app.
-///
-/// ```dart
-/// MegaApp(
-///   home: MegaScaffold(
-///     headerBar: MegaHeaderBar(
-///       center: const Text('MegaApp Theme'),
-///     ),
-///   ),
-/// )
-/// ```
-/// {@end-tool}
 class MegaApp extends StatefulWidget {
-  /// Creates a MegaApp.
-  ///
-  /// At least one of [home], [routes], [onGenerateRoute], or [builder] must be
-  /// non-null. If only [routes] is given, it must include an entry for the
-  /// [Navigator.defaultRouteName] (`/`), since that is the route used when the
-  /// application is launched with an intent that specifies an otherwise
-  /// unsupported route.
-  ///
-  /// This class creates an instance of [WidgetsApp].
-  ///
-  /// The boolean arguments, [routes], and [navigatorObservers], must not be null.
   const MegaApp({
     Key? key,
     this.navigatorKey,
@@ -132,7 +41,6 @@ class MegaApp extends StatefulWidget {
         backButtonDispatcher = null,
         super(key: key);
 
-  /// Creates a [MegaApp] that uses the [Router] instead of a [Navigator].
   const MegaApp.router({
     Key? key,
     this.routeInformationProvider,
@@ -168,215 +76,83 @@ class MegaApp extends StatefulWidget {
         initialRoute = null,
         super(key: key);
 
-  /// {@macro flutter.widgets.widgetsApp.navigatorKey}
   final GlobalKey<NavigatorState>? navigatorKey;
 
-  /// {@macro flutter.widgets.widgetsApp.home}
   final Widget? home;
 
-  /// The application's top-level routing table.
-  ///
-  /// When a named route is pushed with [Navigator.pushNamed], the route name is
-  /// looked up in this map. If the name is present, the associated
-  /// [widgets.WidgetBuilder] is used to construct a [MegaPageRoute] that
-  /// performs an appropriate transition, including [Hero] animations, to the
-  /// new route.
-  ///
-  /// {@macro flutter.widgets.widgetsApp.routes}
   final Map<String, WidgetBuilder>? routes;
 
-  /// {@macro flutter.widgets.widgetsApp.initialRoute}
   final String? initialRoute;
 
-  /// {@macro flutter.widgets.widgetsApp.onGenerateRoute}
   final RouteFactory? onGenerateRoute;
 
-  /// {@macro flutter.widgets.widgetsApp.onGenerateInitialRoutes}
   final InitialRouteListFactory? onGenerateInitialRoutes;
 
-  /// {@macro flutter.widgets.widgetsApp.onUnknownRoute}
   final RouteFactory? onUnknownRoute;
 
-  /// {@macro flutter.widgets.widgetsApp.navigatorObservers}
   final List<NavigatorObserver>? navigatorObservers;
 
-  /// {@macro flutter.widgets.widgetsApp.routeInformationProvider}
   final RouteInformationProvider? routeInformationProvider;
 
-  /// {@macro flutter.widgets.widgetsApp.routeInformationParser}
   final RouteInformationParser<Object>? routeInformationParser;
 
-  /// {@macro flutter.widgets.widgetsApp.routerDelegate}
   final RouterDelegate<Object>? routerDelegate;
 
-  /// {@macro flutter.widgets.widgetsApp.backButtonDispatcher}
   final BackButtonDispatcher? backButtonDispatcher;
 
-  /// {@macro flutter.widgets.widgetsApp.builder}
   final TransitionBuilder? builder;
 
-  /// {@macro flutter.widgets.widgetsApp.title}
-  ///
-  /// This value is passed unmodified to [WidgetsApp.title].
   final String title;
 
-  /// {@macro flutter.widgets.widgetsApp.onGenerateTitle}
-  ///
-  /// This value is passed unmodified to [WidgetsApp.onGenerateTitle].
   final GenerateAppTitle? onGenerateTitle;
 
-  /// {@macro flutter.widgets.widgetsApp.color}
   final Color? color;
 
-  /// {@macro flutter.widgets.widgetsApp.locale}
   final Locale? locale;
 
-  /// {@macro flutter.widgets.widgetsApp.localizationsDelegates}
   final Iterable<LocalizationsDelegate<dynamic>>? localizationsDelegates;
 
-  /// {@macro flutter.widgets.widgetsApp.localeListResolutionCallback}
-  ///
-  /// This callback is passed along to the [WidgetsApp] built by this widget.
   final LocaleListResolutionCallback? localeListResolutionCallback;
 
-  /// {@macro flutter.widgets.LocaleResolutionCallback}
-  ///
-  /// This callback is passed along to the [WidgetsApp] built by this widget.
   final LocaleResolutionCallback? localeResolutionCallback;
 
-  /// {@macro flutter.widgets.widgetsApp.supportedLocales}
-  ///
-  /// It is passed along unmodified to the [WidgetsApp] built by this widget.
   final Iterable<Locale> supportedLocales;
 
-  /// Turns on a performance overlay.
-  ///
-  /// See also:
-  ///
-  ///  * <https://flutter.dev/debugging/#performanceoverlay>
   final bool showPerformanceOverlay;
 
-  /// Turns on checkerboarding of raster cache images.
   final bool checkerboardRasterCacheImages;
 
-  /// Turns on checkerboarding of layers rendered to offscreen bitmaps.
   final bool checkerboardOffscreenLayers;
 
-  /// Turns on an overlay that shows the accessibility information
-  /// reported by the framework.
   final bool showSemanticsDebugger;
 
-  /// {@macro flutter.widgets.widgetsApp.debugShowCheckedModeBanner}
   final bool debugShowCheckedModeBanner;
 
-  /// {@macro flutter.widgets.widgetsApp.shortcuts}
-  /// {@tool snippet}
-  /// This example shows how to add a single shortcut for
-  /// [LogicalKeyboardKey.select] to the default shortcuts without needing to
-  /// add your own [Shortcuts] widget.
-  ///
-  /// Alternatively, you could insert a [Shortcuts] widget with just the mapping
-  /// you want to add between the [WidgetsApp] and its child and get the same
-  /// effect.
-  ///
-  /// ```dart
-  /// Widget build(BuildContext context) {
-  ///   return WidgetsApp(
-  ///     shortcuts: <ShortcutActivator, Intent>{
-  ///       ... WidgetsApp.defaultShortcuts,
-  ///       const SingleActivator(LogicalKeyboardKey.select): const ActivateIntent(),
-  ///     },
-  ///     color: const Color(0xFFFF0000),
-  ///     builder: (BuildContext context, Widget? child) {
-  ///       return const Placeholder();
-  ///     },
-  ///   );
-  /// }
-  /// ```
-  /// {@end-tool}
-  /// {@macro flutter.widgets.widgetsApp.shortcuts.seeAlso}
   final Map<ShortcutActivator, Intent>? shortcuts;
 
-  /// {@macro flutter.widgets.widgetsApp.actions}
-  /// {@tool snippet}
-  /// This example shows how to add a single action handling an
-  /// [ActivateAction] to the default actions without needing to
-  /// add your own [Actions] widget.
-  ///
-  /// Alternatively, you could insert a [Actions] widget with just the mapping
-  /// you want to add between the [WidgetsApp] and its child and get the same
-  /// effect.
-  ///
-  /// ```dart
-  /// Widget build(BuildContext context) {
-  ///   return WidgetsApp(
-  ///     actions: <Type, Action<Intent>>{
-  ///       ... WidgetsApp.defaultActions,
-  ///       ActivateAction: CallbackAction<Intent>(
-  ///         onInvoke: (Intent intent) {
-  ///           // Do something here...
-  ///           return null;
-  ///         },
-  ///       ),
-  ///     },
-  ///     color: const Color(0xFFFF0000),
-  ///     builder: (BuildContext context, Widget? child) {
-  ///       return const Placeholder();
-  ///     },
-  ///   );
-  /// }
-  /// ```
-  /// {@end-tool}
-  /// {@macro flutter.widgets.widgetsApp.actions.seeAlso}
   final Map<Type, Action<Intent>>? actions;
 
-  /// {@macro flutter.widgets.widgetsApp.restorationScopeId}
   final String? restorationScopeId;
 
-  /// {@macro flutter.material.materialApp.scrollBehavior}
-  ///
-  /// When null, defaults to [MegaScrollBehavior].
-  ///
-  /// See also:
-  ///
-  ///  * [ScrollConfiguration], which controls how [Scrollable] widgets behave
-  ///    in a subtree.
   final ScrollBehavior? scrollBehavior;
 
-  /// {@macro flutter.widgets.widgetsApp.useInheritedMediaQuery}
   final bool useInheritedMediaQuery;
 
   @override
   _MegaAppState createState() => _MegaAppState();
 
-  /// The [HeroController] used for Mega page transitions.
-  ///
-  /// Used [MegaApp].
   static HeroController createMegaHeroController() => HeroController();
 }
 
-/// Describes how [Scrollable] widgets behave for [MegaApp]s.
-///
-/// {@macro flutter.widgets.scrollBehavior}
-///
-/// Setting a [MegaScrollBehavior] will result in descendant [Scrollable] widgets
-/// using [BouncingScrollPhysics] by default. No [GlowingOverscrollIndicator] is
-/// applied when using a [MegaScrollBehavior] either, regardless of platform.
-/// When executing on desktop platforms, a [MegaScrollbar] is applied to the child.
-///
-/// See also:
-///
-///  * [ScrollBehavior], the default scrolling behavior extended by this class.
 class MegaScrollBehavior extends ScrollBehavior {
-  /// Creates a MegaScrollBehavior that uses [BouncingScrollPhysics] and
-  /// adds [MegaScrollbar]s on desktop platforms.
   const MegaScrollBehavior();
 
   @override
   Widget buildScrollbar(
-      BuildContext context, Widget child, ScrollableDetails details) {
-    // When modifying this function, consider modifying the implementation in
-    // the base class as well.
+    BuildContext context,
+    Widget child,
+    ScrollableDetails details,
+  ) {
     switch (getPlatform(context)) {
       case TargetPlatform.linux:
       case TargetPlatform.macOS:
@@ -413,11 +189,6 @@ class _MegaAppState extends State<MegaApp> {
     _heroController = MegaApp.createMegaHeroController();
   }
 
-  // Combine the Localizations for Material with the ones contributed
-  // by the localizationsDelegates parameter, if any. Only the first delegate
-  // of a particular LocalizationsDelegate.type is loaded so the
-  // localizationsDelegate parameter can be used to override
-  // _MaterialLocalizationsDelegate.
   Iterable<LocalizationsDelegate<dynamic>> get _localizationsDelegates {
     return <LocalizationsDelegate<dynamic>>[
       if (widget.localizationsDelegates != null)